Save Validation Data - Part 1
The backend has been updated to support keeping track of true/false health status for each condition.

The frontend now properly displays conditions, but with improper logic that needs to be corrected. The frontend now also needs to keep track of updates to the appointment model to make a put call.