Fixed validation page defaults not saving
Validation page will have selected condition checkboxes that default to "true/checked" actually stored in the selectedConditions array to be passed to the database.

This stops the previous buggy behavior where loading a page and saving conditions marked in the database as true would be overwritten back to false. Now the page's truth should always be saved in the database when submitted.